update to make types unsigned ints after disconnect from atomic type

+++ Dice.h
@@ -19,7 +19,7 @@ class Dice {
 public:
     struct result_type {
         std::string report;
-        double result;
+        unsigned int result;
 
         result_type():
         report(),
